transformer/requirements.txt
certifi==2022.9.24
charset-normalizer==2.1.1
contourpy==1.0.6
cycler==0.11.0
editdistance==0.6.1
fonttools==4.38.0
idna==3.4
kiwisolver==1.4.4
matplotlib==3.6.2
numpy==1.23.5
opencv-python==4.6.0.66
packaging==21.3
Pillow==9.3.0
pyflakes==3.0.1
pyparsing==3.0.9
python-dateutil==2.8.2
requests==2.28.1
six==1.16.0
tomli==2.0.1
torch==1.13.0
torchvision==0.14.0
typing_extensions==4.4.0
urllib3==1.26.13
